+++ app.py
@@ -62,8 +62,8 @@ if not MANAGE_DB:
         get_graph_options,
         get_error_and_warnings_div,
     )
-    from layout.imprint import IMPRINT_LAYOUT
-    from layout.privacy import PRIVACY_LAYOUT
+    from layout.imprint import get_imprint_layout
+    from layout.privacy import get_privacy_layout
     from layout.paper import get_paper_layout
 
     app.layout = DEFAULT_LAYOUT
@@ -71,8 +71,8 @@ if not MANAGE_DB:
         [
             DEFAULT_LAYOUT,
             get_layout(app, scenarios=scenario.get_scenarios()),
-            IMPRINT_LAYOUT,
-            PRIVACY_LAYOUT,
+            get_imprint_layout(app),
+            get_privacy_layout(app),
             get_paper_layout(app)
         ]
     )
@@ -85,9 +85,9 @@ if not MANAGE_DB:
 )
 def display_page(pathname):
     if pathname == "/imprint":
-        return IMPRINT_LAYOUT
+        return get_imprint_layout(app)
     elif pathname == "/privacy":
-        return PRIVACY_LAYOUT
+        return get_privacy_layout(app)
     elif pathname == "/paper":
         return get_paper_layout(app)
     else: